requirements.txt
absl-py==0.7.1
alabaster==0.7.12
arrow==0.14.6
asn1crypto==0.24.0
astor==0.8.0
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
backports.csv==1.0.7
backports.functools-lru-cache==1.5
beautifulsoup4==4.7.1
bert-serving-client==1.9.2
bert-serving-server==1.9.2
bibtexparser==1.1.0
bleach==3.1.0
blis==0.2.4
boto==2.49.0
boto3==1.9.182
botocore==1.12.182
bpemb==0.3.0
cffi==1.12.3
chardet==3.0.4
cheroot==6.5.5
CherryPy==18.1.2
Click==7.0
cryptography==2.7
cycler==0.10.0
cymem==2.0.2
decorator==4.4.0
defusedxml==0.6.0
Deprecated==1.2.5
docopt==0.6.2
docutils==0.14
drugstandards==0.5
entrypoints==0.3
feedparser==5.2.1
flair==0.4.2
future==0.17.1
gast==0.2.2
gensim==3.7.3
google-pasta==0.1.7
GPUtil==1.4.0
grpcio==1.22.0
h5py==2.9.0
hyperopt==0.1.2
idna==2.8
imagesize==1.1.0
importlib-metadata==0.18
inflect==2.1.0
ipdb==0.12.1
ipykernel==5.1.1
ipympl==0.2.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
jaraco.functools==2.0
jedi==0.14.1
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
json5==0.8.4
jsonschema==3.0.1
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.2.3
jupyterlab-server==1.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
language-check==1.1
lxml==4.3.4
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mkl-fft==1.0.12
mkl-random==1.0.2
more-itertools==7.1.0
mpld3==0.3
murmurhash==1.0.2
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.3
notebook==6.0.0
num2words==0.5.10
numpy==1.16.4
olefile==0.46
packaging==19.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
git+https://github.com/laasyasree/pattern.git@070f9844a79b655324800382ffea5b8caf2b12bb#egg=pattern
pdfminer.six==20181108
pdfrw==0.4
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
Pint==0.9
plac==0.9.6
plotly==4.0.0
pluggy==0.12.0
portend==2.5
preshed==2.0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
py==1.8.0
pycontractions==2.0.0
pycparser==2.19
pycryptodome==3.8.2
pyemd==0.5.1
Pygments==2.4.2
pymongo==3.8.0
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyrsistent==0.15.3
pytest==5.0.0
python-dateutil==2.8.0
python-docx==0.8.10
python-dotenv==0.10.3
python-Levenshtein==0.12.0
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.0.0
pytz==2019.1
pyzmq==18.0.2
qtconsole==4.5.1
quantulum3==0.7.2
regex==2019.6.8
reportlab==3.5.34
requests==2.22.0
retrying==1.3.3
rst2pdf==0.96
s3transfer==0.2.1
sacremoses==0.0.24
scholarly==0.2.5
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
segtok==1.5.7
semanticscholar==0.1.3
Send2Trash==1.5.0
sentencepiece==0.1.82
six==1.12.0
sklearn==0.0
smart-open==1.8.4
smartypants==2.0.1
snowballstemmer==1.9.0
sortedcontainers==2.1.0
soupsieve==1.9.2
spacy==2.1.6
Sphinx==2.1.2
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-mermaid==0.3.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlitedict==1.6.0
srsly==0.0.7
stemming==1.0.1
strsim==0.0.3
subword-nmt==0.3.6
tabulate==0.8.3
tempora==1.14.1
tensorboard==1.14.0
tensorboardX==1.8
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-hub==0.5.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
textdistance==4.1.3
thinc==7.0.8
torch==1.1.0
torchvision==0.3.0
tornado==6.0.3
tqdm==4.32.2
traitlets==4.3.2
urllib3==1.24.3
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.5.0
wikipedia==1.4.0
wrapt==1.11.2
zc.lockfile==1.4
zipp==0.5.1
